work on wrangle.scanonevar.inputs
preliminarily done with the validation and wrangling portions of scanonevar

wrote lots of tests in the testthat framework, all of which currently pass

one design decision that may need to change is the extreme long-ness of the genoprob.df returned by the wrangling functions.  May want that to the pretty long, but still have the possible alleles as columns (slightly wide), since it will never be used in any longer format than that.  May want to keep it super long, since that's easier to work with.  When I need to use it, it will be filtered down to a small subset, where widening may be easier (there are some name collisions when trying to widen the full genoprob.df)

Signed-off-by: Robert Corty <rcorty@gmail.com>